embeddings: rename package
