+++ embeddings/doc.go
@@ -1,3 +1,3 @@
 // Package embeddings contains a standard interface and implementations
 // for creating numerical representation of textual data
-package embedding
+package embeddings
